Implementate copy constructor, copy assignment in event-related class, modify methods dealing with pointer class member

+++ test/basicPlace.h
@@ -7,7 +7,7 @@
 #include <iostream>
 #include <ctime>
 using namespace std;
-typedef unsigned long UINT;
+typedef unsigned int UINT;
 
 //work as a struct
 class PlaceInfo
